Implementation of basic statistics and user related exercises
The user is now able to view basic statistics of a given subject
related to that user. The statistics shows how every student rated the
exercise.

Each professor will now get a different subject list to be able to
navigate on. The professor gets a list of all the subject he/she teaches