Implementation of iOS segmented control, pieChart statistics and barChart statistics.
iOS segmented control:

Today a segmented control was implemented. This is used in the subject
main page controller to be able to switch between rating statistics and
«time spent» statistics. The possibility to add more segments is open.

PieChart statistics:

Code to view the different resource amounts in a pie chart was
implemented. Now the professor is able to see the percentage amount of
a given resource in a specific exercise.

BarChart Statistics:

This implementation gives the professor the ability to view how many
students that used a given resource the most. This is represented in a
bar graph.