Implementation of 3 more charts and unit tests
3 more charts have been added to the application

1) The first chart is implemented in the subject main page in the
application. This chart is a line chart which shows the percentage of
use of a give resource on a given exercise. All the different resources
are displayed

2) The second chart implemented is a chart which shows how many
students who rated the exercise within a given limit. This is
implemented as a bar graph.

3) The third chart implemented is a chart which shows how much time how
many students used to solve the exercise. This is also a bar graph
where each pillar represents a number of students

Unit test has also been implemented in this working session. The work
of testing the iOS application is still ongoing, and there are
limitations on what´s possible to test and not. There will be
implemented more unit tests