Implementation of the about page of the application. Improvement of the GUI and more functionality for the professor
The about page of the application was implemented, and are able to
download and read pdf documents from the database

Improvement of the GUI to make the iOS application look more similar to
the web application

Implementation of more functionality for the professors. The ability to
register and unregister new subjects is now present.